Refactor errors to separate file
