+++ molpro-dirman/molpro_dirman/errors.py
@@ -1,0 +1,14 @@
+class ProjectSymLinkException(Exception):
+  "General exception during project symlink attempt"
+
+class ProjectSymLinkFailure(ProjectSymLinkException):
+  "Unrecoverable failure during project symlink attempt"
+
+class ProjectSymLinkExists(ProjectSymLinkException):
+  "SymLink Failed because symlink location already exists"
+
+class ProjectAlreadySymLinked(ProjectSymLinkException):
+  "SymLink already exists in the requested configuration"
+
+class ProjectSymlinkedElsewhere(ProjectSymLinkException):
+  "Project is symlinked already, elsewhere in the symlink directory path"
