+++ molpro-dirman/molpro_dirman/sys_write.py
@@ -6,22 +6,12 @@ from pathlib import Path
 
 from .config import Config, symlink_name
 from .sys_read import Project
-
-class ProjectSymLinkException(Exception):
-  "General exception during project symlink attempt"
-
-class ProjectSymLinkFailure(ProjectSymLinkException):
-  "Unrecoverable failure during project symlink attempt"
-
-class ProjectSymLinkExists(ProjectSymLinkException):
-  "SymLink Failed because symlink location already exists"
-
-class ProjectAlreadySymLinked(ProjectSymLinkException):
-  "SymLink already exists in the requested configuration"
-
-class ProjectSymlinkedElsewhere(ProjectSymLinkException):
-  "Project is symlinked already, elsewhere in the symlink directory path"
-
+from .errors import (
+  ProjectSymLinkExists, 
+  ProjectSymLinkFailure, 
+  ProjectSymlinkedElsewhere, 
+  ProjectAlreadySymLinked
+)
 
 def delete_symlink(path: Path, mpdman_only=True) -> Path:
   "Safe method to delete only symlinks. Also can perform check to ensure it is an mpdman-created symlink"
